+++ qbraid/transpiler/conversions/qiskit/qiskit_extras.py
@@ -28,11 +28,13 @@ from qbraid.transpiler.annotations import requires_extras
 qiskit_braket_provider = LazyLoader("qiskit_braket_provider", globals(), "qiskit_braket_provider")
 qbraid_qir_qiskit = LazyLoader("qbraid_qir_qiskit", globals(), "qbraid_qir.qiskit")
 qiskit_ionq = LazyLoader("qiskit_ionq", globals(), "qiskit_ionq")
+qat_interop_qiskit = LazyLoader("qat_interop_qiskit", globals(), "qat.interop.qiskit")
 
 if TYPE_CHECKING:
     import braket.circuits
     import pyqir
     import qiskit.circuit
+    from qat.core.wrappers.circuit import Circuit as QatCircuit  # type: ignore
 
     import qbraid.programs
 
@@ -92,11 +94,8 @@ def qiskit_to_ionq(circuit: qiskit.circuit.QuantumCircuit, **kwargs) -> qbraid.p
     }
 
 
-qat_interop_qiskit = LazyLoader("qat_interop_qiskit", globals(), "qat.interop.qiskit")
-
-
 @requires_extras("qat.interop.qiskit", "qiskit")
-def qiskit_to_qat(circuit: "qiskit.circuit.QuantumCircuit"):
+def qiskit_to_qat(circuit: qiskit.circuit.QuantumCircuit) -> QatCircuit:
     """Returns a qat.core.wrappers.circuit.Circuit equivalent to the input Qiskit circuit.
 
     Args:
@@ -109,7 +108,7 @@ def qiskit_to_qat(circuit: "qiskit.circuit.QuantumCircuit"):
 
 
 @requires_extras("qat.interop.qiskit", "qiskit")
-def qat_to_qiskit(circuit):
+def qat_to_qiskit(circuit: QatCircuit) -> qiskit.circuit.QuantumCircuit:
     """Returns a Qiskit QuantumCircuit equivalent to the input qat circuit.
 
     Args: